+++ packages/fighting-design/button-group/src/props.ts
@@ -7,11 +7,11 @@ export const Props = {
    * 按钮尺寸
    *
    * @values large middle small mini
-   * @defaultValue middle
+   * @defaultValue null
    */
   size: {
     type: String as PropType<FightingSize>,
-    default: (): FightingSize => 'middle',
+    default: (): null => null,
     validator: (val: FightingSize): boolean => {
       return (['large', 'middle', 'small', 'mini'] as const).includes(val)
     }
@@ -33,6 +33,4 @@ export const Props = {
 
 export type ButtonGroupProps = ExtractPropTypes<typeof Props>
 
-export const BUTTON_GROUP_PROPS_KEY: InjectionKey<ButtonGroupProps> = Symbol(
-  'button-group-props-key'
-)
+export const BUTTON_GROUP_PROPS_KEY: InjectionKey<ButtonGroupProps> = Symbol('button-group-props-key')