+++ packages/fighting-design/button/src/props.ts
@@ -42,11 +42,11 @@ export const Props = {
    * 按钮尺寸
    *
    * @values large middle small mini
-   * @defaultValue middle
+   * @defaultValue null
    */
   size: {
     type: String as PropType<FightingSize>,
-    default: (): FightingSize => 'middle',
+    default: (): null => null,
     validator: (val: FightingSize): boolean => {
       return (['large', 'middle', 'small', 'mini'] as const).includes(val)
     }
@@ -104,7 +104,7 @@ export const Props = {
    * 按钮的类型（非自定义按钮颜色时有效）
    *
    * @values default primary success danger warning
-   * @defaultValue default
+   * @defaultValue null
    */
   type: {
     type: String as PropType<FightingType>,